requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.4
    # via requests
idna==3.2
    # via requests
python-dateutil==2.8.2
    # via -r requirements.in
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==1.26.6
    # via requests
